Code refactor User can signup and login User can view all products and add products User can add users User can logout on those pages
