+++ UI/.cache/04/6816ed7fb68a9c6b9e2c2068ea52a4.json
@@ -1,0 +1,1 @@
+{"id":"../static/css/products.css","dependencies":[{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/static/css/products.css","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"css":".search\n{\n    border: 2px solid #CF5C3F;\n    border-radius: 12px;\n    margin: 50px  250px;\n    padding: 0;\n    overflow: hidden;\n    width: 60%\n}\n.welcome a {\n    text-decoration: none !important;\n    padding-top: 4px;\n}\n.search input[type=\"text\"]\n{\n    border: 0px;\n    width: 67%;\n    padding: 15px;\n}\n.modal input[type=number] {\n    width: 100%;\n    padding: 12px;\n    border: 1px solid #ccc;\n    border-radius: 4px;\n    resize: vertical;\n}\n.search input[type=\"text\"]:focus\n{\n    outline: 0;\n}\n\n.search input[type=\"submit\"]\n{\n    border: 0px;\n    margin-top: 0px;\n    margin-bottom: 0px;\n    background: none;\n    background-color: #CF5C3F;\n    color: #fff;\n    float: right;\n    padding: 15px;\n    border-radius: 4px;\n    cursor:pointer;\n}\n\n.welcome {\n    background-color: white;\n    border-radius: 4px;\n}\n\n\n.welcome ul {\n    margin-top: 20px;\n    display: grid;\n    grid-template-columns: repeat(4, 1fr);\n    grid-gap: 10px;\n \n  }\n\n.welcome ul li {\n    color: rgb(37, 37, 99);\n    padding: 10px 8px;\n    list-style: none;\n    font-size: 1.2em;\n\n}\n.dets {\n    color: cadetblue;\n    padding: 10px 10px;\n}\n.welcome input[type=submit]{\n    background-color: #9E6F19 !important;\n    color: #363647;\n    font-size: .9em;\n}\n@media screen and (max-width: 812px) {\n    ul {\n    -moz-column-count: 2;\n    -moz-column-gap: 10px;\n    -webkit-column-count: 2;\n    -webkit-column-gap: 10px;\n    column-count: 2;\n    column-gap: 10px;\n}\n}\n@media screen and (max-width: 1024px) {\n    .search {\n        margin-top: 100px;\n        width: 70%;\n        margin-left: 100px;\n    }\n    \n}\n@media screen and (max-width: 812px) {\n    .search input[type=\"text\"]\n{\n    border: 0px;\n    width: 60%;\n    padding: 14px;\n}\n\n.search input[type=\"submit\"]{\n    padding: 6px;\n}\n}\n","js":"var reloadCSS = require('_css_loader');\n\nmodule.hot.dispose(reloadCSS);\nmodule.hot.accept(reloadCSS);"},"hash":"dda211876fab3bbc87da30c81b8de850","cacheData":{"env":{}}}
