+++ UI/.cache/09/916802429571a743dd0ab4c7f879f7.json
@@ -1,0 +1,1 @@
+{"id":"../static/css/stock.css","dependencies":[{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/static/css/stock.css","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"css":"h1 {\n\tcolor: #0099C4;\n\tfont-style: italic;\n}\n.status{\n    font-style: italic;\n    font-size: 0.7em;\n}\n.stock ul li {\n    color: rgb(37, 37, 99);\n    font-size: 1.3em;\n    width: 85%;\n    border-radius: 4px;\n    margin-top: 15px;\n    background-color: rgb(180, 172, 172);\n    padding: 11px 20px;\n    list-style: none;\n    float: left;\n}\n .status h3,h4 {\n \tpadding-top: 5px;\n \tmargin: 0;\n }\n.pcontainer {\n    border-radius: 5px;\n    background-color: #f2f2f2;\n    margin: 50px 350px;\n    padding: 10px 30px;\n    overflow: hidden;\n    width: 35%;\n}\n.stock input[type=submit]{\n\tmargin-left: 10px;\n\n}\n.btns {\n\tdisplay: inline-block !important;\n}\n.del input[type=submit]{\n\tbackground-color: #CF5C3F !important;\n}\n@media screen and (max-width: 1024px) {\n    .pcontainer {\n        width: 60%;\n        margin-left:40px;\n        padding: 4px;\n    }\n}\n\n@media screen and (max-width: 812px) {\n    .pcontainer {\n        margin-left: 40px;\n        width: 70%;\n    }\n} \n","js":"var reloadCSS = require('_css_loader');\n\nmodule.hot.dispose(reloadCSS);\nmodule.hot.accept(reloadCSS);"},"hash":"5565f42034035741d53e7e66d3f7c891","cacheData":{"env":{}}}
