+++ UI/.cache/1b/66117ab94d9826a891b95c5a608cce.json
@@ -1,0 +1,1 @@
+{"id":"stock.html","dependencies":[{"name":"./../static/css/styles.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/styles.css","parent":"/home/kwanj/andela/store_manager/UI/html/stock.html"},{"name":"./../static/css/stock.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/stock.css","parent":"/home/kwanj/andela/store_manager/UI/html/stock.html"},{"name":"./../static/css/modal.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/modal.css","parent":"/home/kwanj/andela/store_manager/UI/html/stock.html"},{"name":"./../static/images/smlogo.Uy","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/images/smlogo.Uy","parent":"/home/kwanj/andela/store_manager/UI/html/stock.html"},{"name":"./index.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/index.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/stock.html"},{"name":"./stock.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/stock.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/stock.html"},{"name":"./sales.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/sales.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/stock.html"},{"name":"./addattendant.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/addattendant.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/stock.html"},{"name":"./products.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/products.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/stock.html"},{"name":"./login.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/login.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/stock.html"},{"name":"./../static/js/script.js","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/js/script.js","parent":"/home/kwanj/andela/store_manager/UI/html/stock.html"},{"name":"./../static/js/admin_products.js","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/js/admin_products.js","parent":"/home/kwanj/andela/store_manager/UI/html/stock.html"},{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/html/stock.html","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"html":"<!DOCTYPE html>\n<html>\n\n<head>\n    <title>StoreManager</title>\n    <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/d48f55fd66d30608c20fd98c2a12c442.css\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/1b839dbc7a56e5666b3a4770dfe64253.css\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/537c8bc975df84080ecb987429aefeb9.css\">\n</head>\n\n<body>\n    <nav>\n        <ul>\n            <li>\n                <img src=\"/142197def99e8251e87d31fa63e3de82.Uy\" class=\"img-responsive\" alt=\"+StoreManager\">\n            </li>\n            <li><a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\">Home</a></li>\n            <li><a href=\"/556d5103a27286054615ccc0f2cd8db0.html\" class=\"active\">Stock</a></li>\n            <li><a href=\"/8a2c4c2a3fbcbbf837946bec26193b87.html\">Sales</a></li>\n            <li><a href=\"/3b6aabd84370eb7913a550f81ef6a0ca.html\">Add User</a></li>\n            <li><a href=\"/90e10f0ec99bb0fec57ed3d08ff9c0de.html\">Attendant Site</a></li>\n            <li><a href=\"/99bad02247f3eae83c76aa3fe371b715.html\" class=\"login\">Logout</a></li>\n        </ul>\n    </nav>\n    \n        <div id=\"addtocart\" class=\"modal\" style=\"display:none\">\n            <form id=\"newproduct\">\n                <div class=\"col-75\"><label>Product Name</label></div>\n                <div class=\"col-75\"><input id=\"name\" type=\"text\" placeholder=\"name of product..\"></div>\n                <div class=\"col-75\"><label>Number_Of_Items</label></div>\n                <div class=\"col-75\"><input id=\"number\" type=\"text\" placeholder=\"20 \"></div>\n                <div class=\"col-75\"><label>Price KSH</label></div>\n                <div class=\"col-75\"><input id=\"price\" type=\"text\" placeholder=\"200\"></div>\n                <p id=\"output\"></p>\n                <a href=\"#\"><input type=\"submit\" value=\"Done\"><br></a>\n            <form>\n        </form></form></div>\n    \n    <div class=\"pcontainer\">\n        <h1>Available Products</h1>\n        <a href=\"#\" onclick=\"adc()\"><input style=\"background-color: #0BC474\" type=\"submit\" value=\"AddProduct\"></a>\n        <div class=\"stock\">\n            <ul>\n                <li>\n                    <h4>Softcare HC</h4>\n                    <div class=\"status\">\n                        <h3>Category: Diaper</h3>\n                        <h3>Number: 20Bales</h3>\n                        <h3>Min_Number: 3Bales</h3>\n                        <h3>Price: @KSH 600</h3>\n                    </div>\n                </li>\n                <div class=\"btns\">\n                    <a href=\"#\" onclick=\"adc()\"><input type=\"submit\" value=\"Update\"></a>\n                    <a href=\"#\" onclick=\"del()\" class=\"del\"><input type=\"submit\" value=\"Delete\"></a>\n                </div><br>\n                \n            </ul>\n        </div>\n    </div>\n\n    <footer class=\"footercentered\">\n        <p class=\"footer-company-motto\">Sit back.StoreManger will handle it.</p>\n        <p class=\"footer-links\">\n            <a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\">Home</a> ·\n            <a id=\"bpopup\" onclick=\"b_show()\">Blog</a>   \n        </p>\n        <p class=\"footer-company-name\">Store Manager.inc &copy; 2018</p>\n    </footer>\n    <script type=\"text/javascript\" src=\"/d8446085e9c229f9f215fdb2af7e1e9c.js\"></script>\n    <script type=\"text/javascript\" src=\"/52ab6492fb7e06b6b3ce281635920297.js\"></script>\n</body>\n\n</html>"},"hash":"7d308348bdc8d56be1c0e9432cdc9c70","cacheData":{"env":{}}}
