+++ UI/.cache/1d/90ed6d009498e550ccb51ddfb1fef1.json
@@ -1,0 +1,1 @@
+{"id":"../static/css/profile.css","dependencies":[{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/static/css/profile.css","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"css":".pcontainer {\n    border-radius: 5px;\n    background-color: #f2f2f2;\n    margin: 50px 350px;\n    padding: 10px 30px;\n    overflow: hidden;\n    width: 35%;\n}\n\n\n.photo {\n    margin: 10px 120px;\n\n}\n.photo img {\n    border-radius: 50px;\n    max-height: 250px;\n}\n\nul {\n    background: #0A0908;\n    border-radius: 4px;\n    display: flex;\n    flex-direction: row;\n    overflow: hidden;\n    margin: 0;\n    padding: 0;\n}\nul li {\n    list-style: none;\n    flex-grow: 1;\n    float: left;\n    text-align: center;\n    width: 16.6667%;\n    /* fallback for non-calc() browsers */\n    width: calc(100% / 6);\n}\nul li a {\n    display: block;\n    text-decoration: none;\n    color: #F2F4F3;\n    padding: 15px 0;\n}\nul li a:hover {\n    background-color: #292D28;\n    color: ivory;\n}\nh6 {\n    color: cadetblue;\n    margin-left: 40px;\n    font-family: sans-serif;\n    font-size: 1.2em;\n}\n.pinfo {\n    color: cadetblue;\n}\nh4 {\n    padding-bottom: 10px;\n    padding-top: 10px;\n    color: cadetblue;\n}\ninput[type=file] {\n    \n    background-color: gainsboro;\n    border-radius: 4px;\n    cursor: pointer;\n}\n.pform label {\n    color: #0D89C6;\n    padding-bottom: 20px;\n}\n\n @media screen and (max-width: 1024px) {\n    .pcontainer {\n        width: 60%;\n        margin-left:40px;\n        padding: 4px;\n    }\n    .photo {\n        margin: 5px 5px;\n    }\n}\n\n@media screen and (max-width: 812px) {\n    .pcontainer {\n        margin-left: 40px;\n        width: 70%;\n    }\n} \n","js":"var reloadCSS = require('_css_loader');\n\nmodule.hot.dispose(reloadCSS);\nmodule.hot.accept(reloadCSS);"},"hash":"55c35c065870a68112bc86418dc71627","cacheData":{"env":{}}}
