+++ UI/.cache/35/0b5a1e4332d6a9621634c084d9d8d8.json
@@ -1,0 +1,1 @@
+{"id":"../static/css/modal.css","dependencies":[{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/static/css/modal.css","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"css":".modal {\n    border-radius: 5px;\n    background-color: #f2f2f2;\n    margin: 50px 250px;\n    padding: 10px 30px;\n    overflow: hidden;\n    width: 25%;\n}\n\ninput[type=text] {\n    width: 100%;\n    padding: 12px;\n    border: 1px solid #ccc;\n    border-radius: 4px;\n    resize: vertical;\n}\n.col-25 {\n  width: 25%;\n  margin-top: 6px;\n}\n.col-75 {\n  width: 75%;\n  margin-top: 6px;\n}\n@media screen and (max-width: 1024px) {\n    .modal {\n        width: 70%;\n        margin-left: 100px;\n    }\n}\n@media screen and (max-width: 812px) {\n    .modal {\n        margin-left: 40px;\n    }\n} \n","js":"var reloadCSS = require('_css_loader');\n\nmodule.hot.dispose(reloadCSS);\nmodule.hot.accept(reloadCSS);"},"hash":"f1676966b1a16eafc9bed9183dcd4e3e","cacheData":{"env":{}}}
