+++ UI/.cache/38/f66dcd7ff009299b6141f210133906.json
@@ -1,0 +1,1 @@
+{"id":"sales.html","dependencies":[{"name":"./../static/css/styles.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/styles.css","parent":"/home/kwanj/andela/store_manager/UI/html/sales.html"},{"name":"./../static/css/salesadmin.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/salesadmin.css","parent":"/home/kwanj/andela/store_manager/UI/html/sales.html"},{"name":"./../static/images/smlogo.Uy","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/images/smlogo.Uy","parent":"/home/kwanj/andela/store_manager/UI/html/sales.html"},{"name":"./index.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/index.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/sales.html"},{"name":"./stock.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/stock.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/sales.html"},{"name":"./addattendant.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/addattendant.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/sales.html"},{"name":"./products.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/products.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/sales.html"},{"name":"./login.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/login.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/sales.html"},{"name":"./sales2.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/sales2.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/sales.html"},{"name":"./../static/js/script.js","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/js/script.js","parent":"/home/kwanj/andela/store_manager/UI/html/sales.html"}],"generated":{"html":"<!DOCTYPE html>\n<html>\n\n<head>\n    <title>StoreManager</title>\n    <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/d48f55fd66d30608c20fd98c2a12c442.css\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/02dc7c8d371e4643f688f54d2f9e3571.css\">\n</head>\n\n<body>\n    <nav>\n        <ul>\n            <li>\n                <img src=\"/142197def99e8251e87d31fa63e3de82.Uy\" class=\"img-responsive\" alt=\"+StoreManager\">\n            </li>\n            <li><a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\">Home</a></li>\n            <li><a href=\"/556d5103a27286054615ccc0f2cd8db0.html\">Stock</a></li>\n            <li><a href=\"#\" class=\"active\">Sales</a></li>\n            <li><a href=\"/3b6aabd84370eb7913a550f81ef6a0ca.html\">Add User</a></li>\n            <li><a href=\"/90e10f0ec99bb0fec57ed3d08ff9c0de.html\">Attendant Site</a></li>\n            <li><a href=\"/99bad02247f3eae83c76aa3fe371b715.html\" class=\"login\">Logout</a></li>\n        </ul>\n    </nav>\n     <div class=\"welcome\">\n         <h1>Attendant Sale Records</h1>\n        <ul>\n            <li> <strong>Kelvin</strong><br>Email:zeuskay@gmail.com </li>\n            <a href=\"/ce47e3a4d5b737d8a1f2bd30462eceb3.html\"><input type=\"submit\" value=\"Details\"><br></a>\n            <li> <strong>Athena</strong><br>Email:athenakay@gmail.com </li>\n            <a href=\"/ce47e3a4d5b737d8a1f2bd30462eceb3.html\"><input type=\"submit\" value=\"Details\"><br></a>\n            <li> <strong>Jay</strong><br>Email:jay@gmail.com </li>\n            <a href=\"/ce47e3a4d5b737d8a1f2bd30462eceb3.html\"><input type=\"submit\" value=\"Details\"><br></a>\n            <li> <strong>Ptah</strong><br>Email:ptah@gmail.com </li>\n            <a href=\"/ce47e3a4d5b737d8a1f2bd30462eceb3.html\"><input type=\"submit\" value=\"Details\"><br></a>\n            <li> <strong>Kwanj</strong><br>Email:zeuskay@gmail.com </li>\n            <a href=\"/ce47e3a4d5b737d8a1f2bd30462eceb3.html\"><input type=\"submit\" value=\"Details\"><br></a>\n            <li> <strong>Cuhi</strong><br>Email:cuhi@gmail.com </li>\n            <a href=\"/ce47e3a4d5b737d8a1f2bd30462eceb3.html\"><input type=\"submit\" value=\"Details\"><br></a>\n        \n        </ul>\n    </div>\n     <footer class=\"footercentered\">\n        <p class=\"footer-company-motto\">Sit back.StoreManger will handle it.</p>\n        <p class=\"footer-links\">\n            <a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\">Home</a> ·\n            <a id=\"bpopup\" onclick=\"b_show()\">Blog</a>   \n        </p>\n        <p class=\"footer-company-name\">Store Manager.inc &copy; 2018</p>\n    </footer>\n     <script type=\"text/javascript\" src=\"/d8446085e9c229f9f215fdb2af7e1e9c.js\"></script>\n</body>\n</html>\n"},"hash":"343d6bb56396cea40ad059b594c8ba33","cacheData":{}}
