+++ UI/.cache/3d/80d4076402314f9bc279680b60f69d.json
@@ -1,0 +1,1 @@
+{"id":"../static/images/athena.jpg","dependencies":[{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"js":"module.exports = \"/689c57a302d6787b159854bae15ea751.jpg\";"},"hash":"2f766a4f86f64cc56736ac788deffba4","cacheData":{"env":{}}}
