+++ UI/.cache/3d/95d333391b6e13ef2cba6fbcc2091f.json
@@ -1,0 +1,1 @@
+{"id":"../node_modules/node-fetch/browser.js","dependencies":[{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822},{"name":"/home/kwanj/andela/store_manager/UI/node_modules/node-fetch/package.json","includedInParent":true,"mtime":1541560761770}],"generated":{"js":"\"use strict\";\n\nmodule.exports = exports = self.fetch; // Needed for TypeScript and Webpack.\n\nexports.default = self.fetch.bind(self);\nexports.Headers = self.Headers;\nexports.Request = self.Request;\nexports.Response = self.Response;","map":{"mappings":[{"generated":{"line":1,"column":0},"source":"../node_modules/node-fetch/browser.js","original":{"line":1,"column":0}},{"name":"module","generated":{"line":3,"column":0},"source":"../node_modules/node-fetch/browser.js","original":{"line":3,"column":0}},{"generated":{"line":3,"column":6},"source":"../node_modules/node-fetch/browser.js","original":{"line":3,"column":6}},{"name":"exports","generated":{"line":3,"column":7},"source":"../node_modules/node-fetch/browser.js","original":{"line":3,"column":7}},{"generated":{"line":3,"column":14},"source":"../node_modules/node-fetch/browser.js","original":{"line":3,"column":0}},{"name":"exports","generated":{"line":3,"column":17},"source":"../node_modules/node-fetch/browser.js","original":{"line":3,"column":17}},{"generated":{"line":3,"column":24},"source":"../node_modules/node-fetch/browser.js","original":{"line":3,"column":24}},{"name":"self","generated":{"line":3,"column":27},"source":"../node_modules/node-fetch/browser.js","original":{"line":3,"column":27}},{"generated":{"line":3,"column":31},"source":"../node_modules/node-fetch/browser.js","original":{"line":3,"column":31}},{"name":"fetch","generated":{"line":3,"column":32},"source":"../node_modules/node-fetch/browser.js","original":{"line":3,"column":32}},{"generated":{"line":3,"column":37},"source":"../node_modules/node-fetch/browser.js","original":{"line":3,"column":0}},{"generated":{"line":3,"column":39},"source":"../node_modules/node-fetch/browser.js","original":{"line":5,"column":0}},{"name":"exports","generated":{"line":5,"column":0},"source":"../node_modules/node-fetch/browser.js","original":{"line":6,"column":0}},{"generated":{"line":5,"column":7},"source":"../node_modules/node-fetch/browser.js","original":{"line":6,"column":7}},{"name":"default","generated":{"line":5,"column":8},"source":"../node_modules/node-fetch/browser.js","original":{"line":6,"column":8}},{"generated":{"line":5,"column":15},"source":"../node_modules/node-fetch/browser.js","original":{"line":6,"column":0}},{"name":"self","generated":{"line":5,"column":18},"source":"../node_modules/node-fetch/browser.js","original":{"line":6,"column":18}},{"generated":{"line":5,"column":22},"source":"../node_modules/node-fetch/browser.js","original":{"line":6,"column":22}},{"name":"fetch","generated":{"line":5,"column":23},"source":"../node_modules/node-fetch/browser.js","original":{"line":6,"column":23}},{"generated":{"line":5,"column":28},"source":"../node_modules/node-fetch/browser.js","original":{"line":6,"column":18}},{"name":"bind","generated":{"line":5,"column":29},"source":"../node_modules/node-fetch/browser.js","original":{"line":6,"column":29}},{"generated":{"line":5,"column":33},"source":"../node_modules/node-fetch/browser.js","original":{"line":6,"column":18}},{"name":"self","generated":{"line":5,"column":34},"source":"../node_modules/node-fetch/browser.js","original":{"line":6,"column":34}},{"generated":{"line":5,"column":38},"source":"../node_modules/node-fetch/browser.js","original":{"line":6,"column":18}},{"generated":{"line":5,"column":39},"source":"../node_modules/node-fetch/browser.js","original":{"line":6,"column":0}},{"name":"exports","generated":{"line":6,"column":0},"source":"../node_modules/node-fetch/browser.js","original":{"line":8,"column":0}},{"generated":{"line":6,"column":7},"source":"../node_modules/node-fetch/browser.js","original":{"line":8,"column":7}},{"name":"Headers","generated":{"line":6,"column":8},"source":"../node_modules/node-fetch/browser.js","original":{"line":8,"column":8}},{"generated":{"line":6,"column":15},"source":"../node_modules/node-fetch/browser.js","original":{"line":8,"column":0}},{"name":"self","generated":{"line":6,"column":18},"source":"../node_modules/node-fetch/browser.js","original":{"line":8,"column":18}},{"generated":{"line":6,"column":22},"source":"../node_modules/node-fetch/browser.js","original":{"line":8,"column":22}},{"name":"Headers","generated":{"line":6,"column":23},"source":"../node_modules/node-fetch/browser.js","original":{"line":8,"column":23}},{"generated":{"line":6,"column":30},"source":"../node_modules/node-fetch/browser.js","original":{"line":8,"column":0}},{"name":"exports","generated":{"line":7,"column":0},"source":"../node_modules/node-fetch/browser.js","original":{"line":9,"column":0}},{"generated":{"line":7,"column":7},"source":"../node_modules/node-fetch/browser.js","original":{"line":9,"column":7}},{"name":"Request","generated":{"line":7,"column":8},"source":"../node_modules/node-fetch/browser.js","original":{"line":9,"column":8}},{"generated":{"line":7,"column":15},"source":"../node_modules/node-fetch/browser.js","original":{"line":9,"column":0}},{"name":"self","generated":{"line":7,"column":18},"source":"../node_modules/node-fetch/browser.js","original":{"line":9,"column":18}},{"generated":{"line":7,"column":22},"source":"../node_modules/node-fetch/browser.js","original":{"line":9,"column":22}},{"name":"Request","generated":{"line":7,"column":23},"source":"../node_modules/node-fetch/browser.js","original":{"line":9,"column":23}},{"generated":{"line":7,"column":30},"source":"../node_modules/node-fetch/browser.js","original":{"line":9,"column":0}},{"name":"exports","generated":{"line":8,"column":0},"source":"../node_modules/node-fetch/browser.js","original":{"line":10,"column":0}},{"generated":{"line":8,"column":7},"source":"../node_modules/node-fetch/browser.js","original":{"line":10,"column":7}},{"name":"Response","generated":{"line":8,"column":8},"source":"../node_modules/node-fetch/browser.js","original":{"line":10,"column":8}},{"generated":{"line":8,"column":16},"source":"../node_modules/node-fetch/browser.js","original":{"line":10,"column":0}},{"name":"self","generated":{"line":8,"column":19},"source":"../node_modules/node-fetch/browser.js","original":{"line":10,"column":19}},{"generated":{"line":8,"column":23},"source":"../node_modules/node-fetch/browser.js","original":{"line":10,"column":23}},{"name":"Response","generated":{"line":8,"column":24},"source":"../node_modules/node-fetch/browser.js","original":{"line":10,"column":24}},{"generated":{"line":8,"column":32},"source":"../node_modules/node-fetch/browser.js","original":{"line":10,"column":0}}],"sources":{"../node_modules/node-fetch/browser.js":"\"use strict\";\n\nmodule.exports = exports = self.fetch;\n\n// Needed for TypeScript and Webpack.\nexports.default = self.fetch.bind(self);\n\nexports.Headers = self.Headers;\nexports.Request = self.Request;\nexports.Response = self.Response;\n"},"lineCount":null}},"hash":"a6792cb1dacbab552a90ec3d6f75177a","cacheData":{"env":{}}}
