+++ UI/.cache/46/9f5fa1d2e83f4a1ea1f39156d783e4.json
@@ -1,0 +1,1 @@
+{"id":"../static/css/hf.css","dependencies":[{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/static/css/hf.css","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"css":"    /* general body */\n    \nbody {\n    background:#E0E4E5;\n    background-blend-mode: color;\n}\n\nfooter {\n    float: left;\n    position: absolute;\n   bottom:-150px;\n}\n\ninput[type=submit] {\n    margin-top: 10px;\n    background-color: #0099C4;\n    padding: 11px 20px;\n    border: none;\n    border-radius: 4px;\n    cursor: pointer;\n}\n.green input[type=submit]{\n    background-color: #0BC474 !important;\n}\ninput[type=text] {\n    width: 100%;\n    padding: 12px;\n    border: 1px solid #ccc;\n    border-radius: 4px;\n    resize: vertical;\n}\ninput[type=password] {\n    width: 100%;\n    padding: 12px;\n    border: 1px solid #ccc;\n    border-radius: 4px;\n    resize: vertical;\n}\n/* Floating column for labels: 25% width */\n.col-25 {\n  width: 25%;\n  margin-top: 6px;\n}\n/* Floating column for inputs: 75% width */\n.col-75 {\n  width: 75%;\n  margin-top: 6px;\n}\n\n@media screen and (max-width: 1024px) {\n    .col-75 {\n        width: 90%;\n        margin: 0;\n    }\n    h1 {\n        display: none;\n    }\n    .homelogin {\n        margin-left: 50px;\n        width: 50%;\n    }\n}\n\n@media screen and (max-width: 812px) {\n    .welcome, {\n        width: 60%;\n    }\n    h1 {\n        display: none;\n    }\n} \n","js":"var reloadCSS = require('_css_loader');\n\nmodule.hot.dispose(reloadCSS);\nmodule.hot.accept(reloadCSS);"},"hash":"4ce9d22f74779b4c5043344d1e5a6f31","cacheData":{"env":{}}}
