+++ UI/.cache/46/e56b790511e62dec5d8effb09dd9e2.json
@@ -1,0 +1,1 @@
+{"id":"newstore.html","dependencies":[{"name":"./../static/css/styles.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/styles.css","parent":"/home/kwanj/andela/store_manager/UI/html/newstore.html"},{"name":"./../static/css/forms.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/forms.css","parent":"/home/kwanj/andela/store_manager/UI/html/newstore.html"},{"name":"./../static/images/smlogo.Uy","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/images/smlogo.Uy","parent":"/home/kwanj/andela/store_manager/UI/html/newstore.html"},{"name":"./index.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/index.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/newstore.html"},{"name":"./login.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/login.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/newstore.html"},{"name":"./../static/js/signup.js","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/js/signup.js","parent":"/home/kwanj/andela/store_manager/UI/html/newstore.html"},{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/html/newstore.html","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"html":"<!DOCTYPE html>\n<html>\n\n<head>\n    <title>StoreManager</title>\n    <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/d48f55fd66d30608c20fd98c2a12c442.css\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/c87129eb07061dbac962341cc815ffe6.css\">\n</head>\n\n<body>\n    <nav>\n        <ul>\n            <li>\n                <img src=\"/142197def99e8251e87d31fa63e3de82.Uy\" class=\"img-responsive\" alt=\"+StoreManager\">\n            </li>\n            <li><a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\">Home</a></li>\n            <li><a href=\"#\" class=\"active\">NewStore</a></li>\n            <li><a href=\"/99bad02247f3eae83c76aa3fe371b715.html\" class=\"login\">Login</a></li>\n        </ul>\n    </nav>\n    <div class=\"imgcontainer\">\n         <img src=\"/142197def99e8251e87d31fa63e3de82.Uy\" class=\"img-form\" alt=\"+StoreManager\">\n         <h1 style=\"color:#0099C4;\">Create an account for your store</h1>\n    </div>\n    <form id=\"addStore\">\n        <div class=\"welcome\">\n            <div class=\"col-25\"><label>Store Name</label></div>\n            <div class=\"col-75\"><input id=\"storename\" type=\"text\" placeholder=\"The name of your store\"></div>\n            <div class=\"col-25\"><label>Type of Store</label></div>\n            <div class=\"col-75\"><input id=\"category\" type=\"text\" placeholder=\"StoreType eg grocery..\"></div>\n            <div class=\"col-75\"><label>Email Address</label></div>\n            <div class=\"col-75\"><input id=\"email\" type=\"text\" placeholder=\"email@example.com..\"></div>\n            <div class=\"col-25\"><label>Password</label></div>\n            <div class=\"col-75\"><input id=\"password\" type=\"password\" placeholder=\"*******\"></div>\n            <div class=\"col-75\"><label>Confirm Password</label></div>\n            <div class=\"col-75\"><input id=\"confirmpassword\" type=\"password\" placeholder=\"*******\"></div>\n            <p id=\"output\"></p>\n            <a> <input type=\"submit\" value=\"CreateStore\"></a>\n                \n        </div>\n    </form>\n     <footer class=\"footercentered\">\n        <p class=\"footer-company-motto\">Sit back.StoreManger will handle it.</p>\n        <p class=\"footer-links\">\n            <a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\">Home</a> ·\n            <a id=\"bpopup\" onclick=\"b_show()\">Blog</a>   \n        </p>\n        <p class=\"footer-company-name\">Store Manager.inc &copy; 2018</p>\n    </footer>\n    <script type=\"module\" src=\"/6f471a00e22b1a60e1c0eb804d6bba85.js\"></script>\n</body>\n</html>\n"},"hash":"fc7e5822257d73898d8f3d1f3624f34c","cacheData":{"env":{}}}
