+++ UI/.cache/53/b7131570cfbb4c2cb6e6d2de3abe85.json
@@ -1,0 +1,1 @@
+{"id":"../static/css/atnt.css","dependencies":[{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/static/css/atnt.css","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"css":"/*\nforms\n*/\n\n\n\n.welcome {\n    margin-left: 400px;\n    margin-top: 80px; \n    border-radius: 5px;\n    background-color: #f2f2f2;\n    color: black;\n    width: 40%;\n}\ninput[type=submit] {\n    margin-top: 10px;\n    background-color: #0099C4;\n    padding: 11px 20px;\n    border: none;\n    border-radius: 4px;\n    cursor: pointer;\n}\n.green input[type=submit]{\n    background-color: #0BC474 !important;\n}\ninput[type=text] {\n    width: 100%;\n    padding: 12px;\n    border: 1px solid #ccc;\n    border-radius: 4px;\n    resize: vertical;\n}\ninput[type=password] {\n    width: 100%;\n    padding: 12px;\n    border: 1px solid #ccc;\n    border-radius: 4px;\n    resize: vertical;\n}\n/* Floating column for labels: 25% width */\n.col-25 {\n  width: 25%;\n  margin-top: 6px;\n}\n/* Floating column for inputs: 75% width */\n.col-75 {\n  width: 75%;\n  margin-top: 6px;\n}\nfooter {\n    left: 0;\n    bottom: 0;\n    position:fixed;\n}\n@media screen and (max-width: 1024px) {\n    .welcome {\n        width: 40%;\n        margin-left: 70px;\n    }\n}\n\n@media screen and (max-width: 812px) {\n    .welcome {\n        margin-left: 50px;\n    }\n} \n","js":"var reloadCSS = require('_css_loader');\n\nmodule.hot.dispose(reloadCSS);\nmodule.hot.accept(reloadCSS);"},"hash":"a65cbfb56bba57cbe07c134783ae630e","cacheData":{"env":{}}}
