+++ UI/.cache/7f/36d43cd942c6d0997ee47eaec06bca.json
@@ -1,0 +1,1 @@
+{"id":"../static/css/forms.css","dependencies":[{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/static/css/forms.css","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"css":"/*\nforms\n*/\n\n\n.imgcontainer {\n    \n    margin: 50px 250px;\n    padding: 10px 30px;\n    overflow: hidden;\n    width: 60%;\n}\n\nh1 {\n    display: inline;\n    color: ivory;\n    font-size: 2em;\n}\n\n.img-form {\n    padding-top: 20px;\n    max-height: 50px;\n}\n\n.welcome {\n    width: 60%;\n    margin-top: 40px; \n    border-radius: 5px;\n    background-color: #f2f2f2;\n    color: black;\n}\ninput[type=submit] {\n    margin-top: 10px;\n    background-color: #0099C4;\n    padding: 11px 20px;\n    border: none;\n    border-radius: 4px;\n    cursor: pointer;\n}\n.green input[type=submit]{\n    background-color: #0BC474 !important;\n}\ninput[type=text] {\n    width: 100%;\n    padding: 12px;\n    border: 1px solid #ccc;\n    border-radius: 4px;\n    resize: vertical;\n}\n\ninput[type=password] {\n    width: 100%;\n    padding: 12px;\n    border: 1px solid #ccc;\n    border-radius: 4px;\n    resize: vertical;\n}\n/* Floating column for labels: 25% width */\n.col-25 {\n  width: 25%;\n  margin-top: 6px;\n}\n/* Floating column for inputs: 75% width */\n.col-75 {\n  width: 95%;\n  margin-top: 6px;\n}\n\n@media screen and (max-width: 1024px) {\n    .col-75 {\n        width: 90%;\n        margin: 0;\n    }\n    img,h1 {\n        display: none;\n    }\n    .welcome, {\n        margin-left: 50px;\n        width: 50%;\n    }\n}\n\n@media screen and (max-width: 812px) {\n    .welcome, {\n        width: 60%;\n    }\n    img,h1 {\n        display: none;\n    }\n} \n","js":"var reloadCSS = require('_css_loader');\n\nmodule.hot.dispose(reloadCSS);\nmodule.hot.accept(reloadCSS);"},"hash":"289440bcc99c61e6da80d348b877ffa9","cacheData":{"env":{}}}
