+++ UI/.cache/94/70ff1e21f92e2f0da728ab3e6a7047.json
@@ -1,0 +1,1 @@
+{"id":"login.html","dependencies":[{"name":"./../static/css/styles.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/styles.css","parent":"/home/kwanj/andela/store_manager/UI/html/login.html"},{"name":"./../static/css/forms.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/forms.css","parent":"/home/kwanj/andela/store_manager/UI/html/login.html"},{"name":"./../static/images/smlogo.Uy","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/images/smlogo.Uy","parent":"/home/kwanj/andela/store_manager/UI/html/login.html"},{"name":"./index.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/index.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/login.html"},{"name":"./newstore.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/newstore.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/login.html"},{"name":"./../static/js/login.js","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/js/login.js","parent":"/home/kwanj/andela/store_manager/UI/html/login.html"},{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/html/login.html","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"html":"<!DOCTYPE html>\n<html>\n\n<head>\n    <title>StoreManager</title>\n    <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/d48f55fd66d30608c20fd98c2a12c442.css\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/c87129eb07061dbac962341cc815ffe6.css\">\n</head>\n\n<body>\n    <nav>\n        <ul>\n            <li>\n                <img src=\"/142197def99e8251e87d31fa63e3de82.Uy\" class=\"img-responsive\" alt=\"+StoreManager\">\n            </li>\n            <li><a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\">Home</a></li>\n            <li><a href=\"/cbca0e5e6b6815dab5570d9c4d73e8bd.html\">NewStore</a></li>\n            <li><a href=\"#\" class=\"active\">Login</a></li>\n        </ul>\n    </nav>\n    <div class=\"imgcontainer\">\n         <img src=\"/142197def99e8251e87d31fa63e3de82.Uy\" class=\"img-form\" alt=\"+StoreManager\">\n         <h1 style=\"color:#0099C4;\">LogIn</h1>\n    </div>\n    <form id=\"login\">\n        <div class=\"welcome\">\n            <div class=\"col-25\"><label>Email Address</label></div>\n            <div class=\"col-75\"><input id=\"email\" type=\"text\" placeholder=\"email@example.com..\"></div>\n            <div class=\"col-25\"><label>Password</label></div>\n            <div class=\"col-75\"><input id=\"password\" type=\"password\" placeholder=\"*******\"></div>\n            <p id=\"output\"></p>\n            <a> <input type=\"submit\" style=\"background-color: #0BC474\" value=\"LogIn\"></a>    \n        </div>\n    </form>\n    <footer class=\"footercentered\">\n        <p class=\"footer-company-motto\">Sit back.StoreManger will handle it.</p>\n        <p class=\"footer-links\">\n            <a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\">Home</a> ·\n            <a id=\"bpopup\" onclick=\"b_show()\">Blog</a>   \n        </p>\n        <p class=\"footer-company-name\">Store Manager.inc &copy; 2018</p>\n    </footer>\n    <script type=\"text/javascript\" src=\"/56435923cbc73eaec475ab953f6337f9.js\"></script>\n</body>\n</html>\n"},"hash":"fee006b6902ad5a0e0c139f2fa64462a","cacheData":{"env":{}}}
