+++ UI/.cache/9a/cfd852bc54484b41a48a0579710602.json
@@ -1,0 +1,1 @@
+{"id":"../static/css/salesadmin.css","dependencies":[{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/static/css/salesadmin.css","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"css":"h1 {\n\tcolor: #1C3041;\n\tfont-style: italic;\n}\n.welcome ul li {\n    color: rgb(37, 37, 99);\n    font-size: 1.2em;\n    width: 65%;\n    border-radius: 4px;\n    margin-top: 15px;\n    background-color: rgb(180, 172, 172);\n    padding: 11px 20px;\n    list-style: none;\n    \n}\n\nli {\n    display: inline-block;\n}\n.welcome {\n    margin-left: 400px;\n    margin-top: 80px; \n    border-radius: 5px;\n    background-color: #f2f2f2;\n    color: black;\n    width: 40%;\n}\n.btns {\n\tdisplay: inline-block !important;\n}\n.del input[type=submit]{\n\tbackground-color: #CF5C3F !important;\n}\n@media screen and (max-width: 1024px) {\n    .welcome {\n        width: 40%;\n        margin-left: 70px;\n    }\n}\n\n@media screen and (max-width: 812px) {\n    .welcome {\n        margin-left: 50px;\n    }\n} \n","js":"var reloadCSS = require('_css_loader');\n\nmodule.hot.dispose(reloadCSS);\nmodule.hot.accept(reloadCSS);"},"hash":"94c1a0c9b421b21c75cf8b1cb8bd2da5","cacheData":{"env":{}}}
