+++ UI/.cache/a3/15ff193d428b299be8811f311b7a4e.json
@@ -1,0 +1,1 @@
+{"id":"addattendant.html","dependencies":[{"name":"./../static/css/styles.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/styles.css","parent":"/home/kwanj/andela/store_manager/UI/html/addattendant.html"},{"name":"./../static/css/atnt.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/atnt.css","parent":"/home/kwanj/andela/store_manager/UI/html/addattendant.html"},{"name":"./../static/images/smlogo.Uy","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/images/smlogo.Uy","parent":"/home/kwanj/andela/store_manager/UI/html/addattendant.html"},{"name":"./index.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/index.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/addattendant.html"},{"name":"./stock.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/stock.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/addattendant.html"},{"name":"./sales.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/sales.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/addattendant.html"},{"name":"./products.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/products.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/addattendant.html"},{"name":"./login.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/login.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/addattendant.html"},{"name":"./../static/js/new_user.js","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/js/new_user.js","parent":"/home/kwanj/andela/store_manager/UI/html/addattendant.html"},{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/html/addattendant.html","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"html":"<!DOCTYPE html>\n<html>\n\n<head>\n    <title>StoreManager</title>\n    <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/d48f55fd66d30608c20fd98c2a12c442.css\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/7d3846d75e46f4cc4ceaf435955004d1.css\">\n</head>\n\n<body>\n    <nav>\n        <ul>\n            <li>\n                <img src=\"/142197def99e8251e87d31fa63e3de82.Uy\" class=\"img-responsive\" alt=\"+StoreManager\">\n            </li>\n            <li><a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\">Home</a></li>\n            <li><a href=\"/556d5103a27286054615ccc0f2cd8db0.html\">Stock</a></li>\n            <li><a href=\"/8a2c4c2a3fbcbbf837946bec26193b87.html\">Sales</a></li>\n            <li><a href=\"#\" class=\"active\">Add User</a></li>\n            <li><a href=\"/90e10f0ec99bb0fec57ed3d08ff9c0de.html\">Attendant Site</a></li>\n            <li><a href=\"/99bad02247f3eae83c76aa3fe371b715.html\" class=\"login\">Logout</a></li>\n        </ul>\n    </nav>\n     <div class=\"welcome\">\n          <div class=\"col-25\"><label>Email Address</label></div>\n          <div class=\"col-75\"><input id=\"email\" type=\"text\" placeholder=\"email@example.com..\"></div>\n          <div class=\"col-25\"><label>Password</label></div>\n          <div class=\"col-75\"><input id=\"password\" type=\"password\" placeholder=\"*******\"></div>\n          <p id=\"output\"></p>\n          <a href=\"#\" onclick=\"addAttendant()\"> <input type=\"submit\" style=\"background-color: #0BC474\" value=\"AddAttendant\"></a>  \n          <a href=\"#\" onclick=\"addAdmin()\"><input type=\"submit\" value=\"MakeAdmin\"></a>    \n     </div>\n    <footer class=\"footercentered\">\n        <p class=\"footer-company-motto\">Sit back.StoreManger will handle it.</p>\n        <p class=\"footer-links\">\n            <a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\">Home</a> ·\n            <a id=\"bpopup\" onclick=\"b_show()\">Blog</a>   \n        </p>\n        <p class=\"footer-company-name\">Store Manager.inc &copy; 2018</p>\n    </footer>\n     <script type=\"text/javascript\" src=\"/b492360af890ec4e57e07a633c9a7119.js\"></script>\n</body>\n</html>\n"},"hash":"fe02c810ff53c767935edf78d24787ee","cacheData":{"env":{}}}
