+++ UI/.cache/a5/d061ee15dbe646191b8a036fb6468e.json
@@ -1,0 +1,1 @@
+{"id":"../static/images/smlogo.Uy","dependencies":[{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"js":"module.exports = \"/142197def99e8251e87d31fa63e3de82.Uy\";"},"hash":"e892911e18af48898a2440991d04d7af","cacheData":{"env":{}}}
