+++ UI/.cache/a6/9a012e4ca3ff672682effb682a7be9.json
@@ -1,0 +1,1 @@
+{"id":"profile.html","dependencies":[{"name":"./../static/css/styles.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/styles.css","parent":"/home/kwanj/andela/store_manager/UI/html/profile.html"},{"name":"./../static/css/profile.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/profile.css","parent":"/home/kwanj/andela/store_manager/UI/html/profile.html"},{"name":"./../static/css/forms.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/forms.css","parent":"/home/kwanj/andela/store_manager/UI/html/profile.html"},{"name":"./../static/images/smlogo.Uy","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/images/smlogo.Uy","parent":"/home/kwanj/andela/store_manager/UI/html/profile.html"},{"name":"./index.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/index.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/profile.html"},{"name":"./products.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/products.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/profile.html"},{"name":"./carts.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/carts.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/profile.html"},{"name":"./stock.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/stock.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/profile.html"},{"name":"./login.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/login.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/profile.html"},{"name":"./mycarts.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/mycarts.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/profile.html"},{"name":"./attendant-sales.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/attendant-sales.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/profile.html"},{"name":"./../static/images/athena.jpg","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/images/athena.jpg","parent":"/home/kwanj/andela/store_manager/UI/html/profile.html"},{"name":"./../static/js/t.js","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/js/t.js","parent":"/home/kwanj/andela/store_manager/UI/html/profile.html"},{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/html/profile.html","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"html":"<!DOCTYPE html>\n<html>\n\n<head>\n    <title>StoreManager</title>\n    <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/d48f55fd66d30608c20fd98c2a12c442.css\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/a895a126a9818d48fc613b7e74482a7f.css\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/c87129eb07061dbac962341cc815ffe6.css\">\n</head>\n\n<body>\n    <nav>\n        <ul>\n            <li>\n                <img src=\"/142197def99e8251e87d31fa63e3de82.Uy\" class=\"img-responsive\" alt=\"+StoreManager\">\n            </li>\n            <li><a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\">Home</a></li>\n            <li><a href=\"/90e10f0ec99bb0fec57ed3d08ff9c0de.html\">Products</a></li>\n            <li><a href=\"/ebbde448b9050d11a7e8190cb2053764.html\">Carts</a></li>\n            <li><a href=\"#\" class=\"active\">Your Profile</a></li>\n            <li><a href=\"/556d5103a27286054615ccc0f2cd8db0.html\">Admin Site</a></li>\n            <li><a href=\"/99bad02247f3eae83c76aa3fe371b715.html\" class=\"login\">Logout</a></li>\n        </ul>\n    </nav>\n\n     <div class=\"pcontainer\" id=\"editp\" style=\"display: none;\">\n            <div class=\"pform\">\n                <h3>Edit Profile</h3>\n                <div class=\"col-25\"> <label>User Name</label></div>\n                <div class=\"col-75\"><input type=\"text\" name=\"\" placeholder=\"athena-kay\"></div>\n                <div class=\"col-75\"><label>Email</label></div>\n                <div class=\"col-75\"><input type=\"text\" name=\"\" placeholder=\"athena.kay@gmail.com\"></div>\n                <div class=\"col-75\"><label>Old Password</label></div>\n                <div class=\"col-75\"> <input type=\"password\" name=\"\" placeholder=\"****\"></div>\n                <div class=\"col-75\"><label>New Password</label></div>\n                <div class=\"col-75\"> <input type=\"password\" name=\"\" placeholder=\"****\"></div>\n                <div class=\"col-75\"><label>Confirm Password</label></div>\n                <div class=\"col-75\"><input type=\"password\" placeholder=\"*******\"></div>\n                <div class=\"col-75\"> <label>Telephone</label></div>\n                <div class=\"col-75\"><input type=\"text\" name=\"\" placeholder=\"0703 xxx xxx\"></div>\n                <div class=\"col-75\"> <label>Change Profile Picture</label></div>\n                <input placeholder=\"choose file\" type=\"file\">\n            </div>\n            <div class=\"absolute\"><input class=\"update\" type=\"submit\" value=\"Update Profile\"></div>\n        </div>\n    \n    <div class=\"pcontainer\">\n        <ul>\n            <li><a href=\"/df3f4d964ad385e734795edbec696eb2.html\">MyCarts</a></li>\n            <li><a href=\"/e405c0b90f7a723e5c79255d92cf4f2e.html\">Sales</a></li>\n            <li><a href=\"#\" onclick=\"editf()\">EditProfile</a></li>\n        </ul>\n        <div class=\"photo\">\n            <img src=\"/689c57a302d6787b159854bae15ea751.jpg\" alt=\"+athena\">\n            <h6> athena-kay</h6>\n        </div>\n        <div class=\"pinfo\">\n            <h4> <strong>Email:</strong>   athena.kay@gmail.com</h4>\n            <h4> <strong>Phone:</strong>   +5567 768 565</h4>\n            <h4> <strong>User Type: Attendant</strong></h4>\n            <h4> Total Amount Sold: KSH 3,457,576</h4>\n        </div>\n    </div>\n\n    <footer class=\"footercentered\">\n        <p class=\"footer-company-motto\">Sit back.StoreManger will handle it.</p>\n        <p class=\"footer-links\">\n            <a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\">Home</a> ·\n            <a id=\"bpopup\" onclick=\"b_show()\">Blog</a>   \n        </p>\n        <p class=\"footer-company-name\">Store Manager.inc &copy; 2018</p>\n    </footer>\n    <script type=\"text/javascript\" src=\"/82766a9dc42d40c4da2306c564a610bc.js\"></script>\n</body>\n\n</html>\n"},"hash":"18f849d8ea078ab0d3b32f449c621ff9","cacheData":{"env":{}}}
