+++ UI/.cache/a7/b6021c8435cf7642b8a6e4251b326e.json
@@ -1,0 +1,1 @@
+{"id":"index.html","dependencies":[{"name":"./../static/css/styles.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/styles.css","parent":"/home/kwanj/andela/store_manager/UI/html/index.html"},{"name":"./../static/css/hf.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/hf.css","parent":"/home/kwanj/andela/store_manager/UI/html/index.html"},{"name":"./../static/images/smlogo.Uy","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/images/smlogo.Uy","parent":"/home/kwanj/andela/store_manager/UI/html/index.html"},{"name":"./index.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/index.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/index.html"},{"name":"./newstore.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/newstore.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/index.html"},{"name":"./login.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/login.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/index.html"},{"name":"./products.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/products.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/index.html"},{"name":"./../static/js/login.js","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/js/login.js","parent":"/home/kwanj/andela/store_manager/UI/html/index.html"},{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/html/index.html","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"html":"<!DOCTYPE html>\n<html>\n\n<head>\n    <title>StoreManager</title>\n    <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/d48f55fd66d30608c20fd98c2a12c442.css\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/b02a68d6a32aaa7072e3f2252a3978d3.css\">\n</head>\n\n<body>\n    <nav>\n        <ul>\n            <li>\n                <img src=\"/142197def99e8251e87d31fa63e3de82.Uy\" class=\"img-responsive\" alt=\"+StoreManager\">\n            </li>\n            <li><a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\" class=\"active\">Home</a></li>\n           \n            <li><a href=\"/cbca0e5e6b6815dab5570d9c4d73e8bd.html\" class=\"cstore\">NewStore</a></li>\n            <li><a href=\"/99bad02247f3eae83c76aa3fe371b715.html\" class=\"login\">Login</a></li>\n        </ul>\n    </nav>\n    <div class=\"footer-push\">\n        <div class=\"helper\">\n            <h1>Store Manager.inc</h1>\n            <p>Store Manager is a web application that helps store owners manage sales and product inventory records. \n                This application is meant for use in a single store.</p>\n                <p>Store Manager is a web application that helps store owners manage sales and product inventory records. \n                        This application is meant for use in a single store.</p>\n            <div class=\"start\">\n                <a href=\"/cbca0e5e6b6815dab5570d9c4d73e8bd.html\"><input type=\"submit\" value=\"Get Started\"></a>\n            </div>\n        </div>\n        <form id=\"login\">\n            <div class=\"homelogin\">\n                <div class=\"col-75\"><label>Email Address</label></div>\n                <div class=\"col-75\"><input id=\"email\" type=\"text\" placeholder=\"email@example.com..\"></div>\n                <div class=\"col-25\"><label>Password</label></div>\n                <div class=\"col-75\"><input id=\"password\" type=\"password\" placeholder=\"*******\"></div>\n                <p id=\"output\"></p>\n                <div class=\"col-75\"><a href=\"/90e10f0ec99bb0fec57ed3d08ff9c0de.html\"> <input type=\"submit\" style=\"background-color: #0BC474\" value=\"LogIn\"></a></div> \n                </div>\n            </form></div>\n        \n    <footer class=\"footercentered\">\n        <p class=\"footer-company-motto\">Sit back.StoreManger will handle it.</p>\n        <p class=\"footer-links\">\n            <a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\">Home</a> ·\n            <a id=\"bpopup\" onclick=\"b_show()\">Blog</a>   \n        </p>\n        <p class=\"footer-company-name\">Store Manager.inc &copy; 2018</p>\n    </footer>\n    <script type=\"text/javascript\" src=\"/56435923cbc73eaec475ab953f6337f9.js\"></script>\n</body>\n\n</html>\n"},"hash":"5687c7e2fcbe422f88a4b33ec16e8f9f","cacheData":{"env":{}}}
