+++ UI/.cache/cf/bf1fdcef58ae3a7361658668f6cc11.json
@@ -1,0 +1,1 @@
+{"id":"../static/css/sales.css","dependencies":[{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/static/css/sales.css","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"css":".sales-container {\n    border-radius: 5px;\n    background-color: #f2f2f2;\n    margin: 50px 150px;\n    padding: 10px 30px;\n    overflow: hidden;\n    width: 75%;\n    font-size: 1.1em;\n    color: #550527;\n}\n\n\ntable {\n    border-collapse: collapse;\n    width: 100%;\n    color: rgb(29, 42, 49);\n}\nth, td {\n    text-align: left;\n    padding: 8px;\n}\ntr:nth-child(even) {\n    background-color: cadetblue;\n}\n @media screen and (max-width: 1024px) {\n    .sales-container {\n        width: 60%;\n        margin-left:40px;\n        padding: 4px;\n    }\n}\n\n@media screen and (max-width: 812px) {\n    .sales-container {\n        margin-left: 40px;\n        width: 70%;\n    }\n} \n","js":"var reloadCSS = require('_css_loader');\n\nmodule.hot.dispose(reloadCSS);\nmodule.hot.accept(reloadCSS);"},"hash":"604f60f540908ea367c27aeb27706ec3","cacheData":{"env":{}}}
