+++ UI/.cache/d8/92b753c1ba4ce6c82709856f422f57.json
@@ -1,0 +1,1 @@
+{"id":"../static/css/styles.css","dependencies":[{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/static/css/styles.css","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"css":"body {\n    background:#E0E4E5;\n    background-blend-mode: color;\n}\n\n\n    \nnav {\n    width: 100%;\n    background: #0A0908;\n    top: 0;\n    left: 4px;\n    right: 16px;\n    position: fixed;\n}\n\nnav ul {\n    display: flex;\n    flex-direction: row;\n    overflow: hidden;\n    margin: 0;\n    padding: 0;\n}\n\nnav ul li {\n    list-style: none;\n    flex-grow: 1;\n    float: left;\n    text-align: center;\n    width: 16.6667%;\n    /* fallback for non-calc() browsers */\n    width: calc(100% / 6);\n}\n\nnav ul li a {\n    display: block;\n    text-decoration: none;\n    color: #F2F4F3;\n    padding: 15px 0;\n}\n\n.login {\n    color: aqua;\n}\n\nnav ul li a:hover {\n    background-color: #292D28;\n    color: ivory;\n}\n\nnav ul li a.active {\n    background-color: #292D28;\n    color: ivory;\n}\n\n.img-responsive {\n    padding-top: 5px;\n    max-height: 35px;\n}\n\n.logohead {\n    color: ivory;\n    display: inline;\n}\n/* ::content */\n\n.helper {\n    color:#0A0908 ;\n    font-size: 1.2em;\n    width: 40%;\n    margin-top: 150px;\n    margin-left: 200px;\n    float: left;\n\n}\n.helper h2,.helper a {\n    padding: 5px 0px;\n}\n\n.vl{\n    display: inline;\n    padding-left: 5px;\n    padding-right: 5px;\n    border-left: 2px solid #0A0908;\n\n}\n\n.welcome {\n    color: honeydew;\n    margin: 50px 250px;\n    padding: 10px 30px;\n    overflow: hidden;\n    width: 60%\n}\n.homelogin {\n    width: 25%;\n    margin-top: 180px;\n    float: right;\n    border-radius: 4px;\n    background-color: #f2f2f2;\n    padding: 30px;\n    margin-right: 120px;\n}\n.welcome h2,.bigcontainer h2,\n.welcome a {\n    padding: 5px 0px;\n}\n\n.welcome p ,{\n    font-size: 1.2em;\n}\n\n.start input {\n    margin-top: 10px !important;\n    background-color: #292D28;\n}\n\ninput[type=submit] {\n    margin-top: 10px;\n    background-color: #0D89C6;\n    padding: 11px 20px;\n    border: none;\n    border-radius: 4px;\n    cursor: pointer;\n}\n @media screen and (max-width: 1024px) {\n    .welcome {\n        width: 70%;\n        margin-left: 100px;\n    }\n    .welcome #about {\n        padding: 0 !important;\n    }\n}\n\n@media screen and (max-width: 812px) {\n    .welcome {\n        margin-left: 40px;\n    }\n    .welcome #about {\n        padding: 0 !important;\n    }\n} \n\n\n\n/*\nfoooter\n*/\n\n.footercentered{\n    background-color: #292c2f;\n    border-radius: 6px;\n    box-shadow: 0 1px 1px 0 rgba(0, 0, 0, 0.12);\n    box-sizing: border-box;\n    width: 100%;\n    text-align: center;\n    font: normal 18px sans-serif;\n    padding: 45px;\n    margin-top: 80px;\n    \n}\n\n.footercentered .footer-company-motto{\n    color:  #8d9093;\n    font-size: 24px;\n    margin: 0;\n}\n\n.footercentered .footer-company-name{\n    color:  #8f9296;\n    font-size: 14px;\n    margin: 0;\n}\n\n.footercentered .footer-links{\n    list-style: none;\n    font-weight: bold;\n    color:  #ffffff;\n    padding: 35px 0 23px;\n    margin: 0;\n}\n\n.footercentered .footer-links a{\n    display:inline-block;\n    text-decoration: none;\n    color: inherit;\n}\n\n@media (max-width: 1024px) {\n\n    .footercentered{\n        padding: 35px;\n    }\n\n    .footercentered .footer-company-motto{\n        font-size: 18px;\n    }\n\n    .footercentered .footer-company-name{\n        font-size: 12px;\n    }\n\n    .footercentered .footer-links{\n        font-size: 14px;\n        padding: 25px 0 20px;\n    }\n\n    .footercentered .footer-links a{\n        line-height: 1.8;\n    }\n}\n","js":"var reloadCSS = require('_css_loader');\n\nmodule.hot.dispose(reloadCSS);\nmodule.hot.accept(reloadCSS);"},"hash":"d2e348f76ec8b05f1b46e98953427351","cacheData":{"env":{}}}
