+++ UI/.cache/f8/0ff323ba8ff2f0502ff6dae062ff17.json
@@ -1,0 +1,1 @@
+{"id":"products.html","dependencies":[{"name":"./../static/css/styles.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/styles.css","parent":"/home/kwanj/andela/store_manager/UI/html/products.html"},{"name":"./../static/css/products.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/products.css","parent":"/home/kwanj/andela/store_manager/UI/html/products.html"},{"name":"./../static/css/modal.css","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/css/modal.css","parent":"/home/kwanj/andela/store_manager/UI/html/products.html"},{"name":"./../static/images/smlogo.Uy","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/images/smlogo.Uy","parent":"/home/kwanj/andela/store_manager/UI/html/products.html"},{"name":"./index.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/index.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/products.html"},{"name":"./carts.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/carts.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/products.html"},{"name":"./profile.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/profile.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/products.html"},{"name":"./stock.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/stock.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/products.html"},{"name":"./login.html","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/html/login.html","entry":true,"parent":"/home/kwanj/andela/store_manager/UI/html/products.html"},{"name":"./../static/js/general_products.js","dynamic":true,"resolved":"/home/kwanj/andela/store_manager/UI/static/js/general_products.js","parent":"/home/kwanj/andela/store_manager/UI/html/products.html"},{"name":"_css_loader","parent":"/home/kwanj/andela/store_manager/UI/html/products.html","resolved":"/home/kwanj/andela/store_manager/UI/node_modules/parcel-bundler/src/builtins/css-loader.js"},{"name":"/home/kwanj/andela/store_manager/UI/package.json","includedInParent":true,"mtime":1541566609822}],"generated":{"html":"<!DOCTYPE html>\n<html>\n\n<head>\n    <title>StoreManager</title>\n    <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/d48f55fd66d30608c20fd98c2a12c442.css\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/d2194c143366e6b744167bc9b9845108.css\">\n    <link rel=\"stylesheet\" type=\"text/css\" href=\"/537c8bc975df84080ecb987429aefeb9.css\">\n</head>\n\n<body>\n    <nav>\n        <ul>\n            <li>\n                <img src=\"/142197def99e8251e87d31fa63e3de82.Uy\" class=\"img-responsive\" alt=\"+StoreManager\">\n            </li>\n            <li><a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\">Home</a></li>\n            <li><a href=\"#\" class=\"active\">Products</a></li>\n            <li><a href=\"/ebbde448b9050d11a7e8190cb2053764.html\">Carts</a></li>\n            <li><a href=\"/5ab5f0cf6ccecdfc06f371694e6b8396.html\">Your Profile</a></li>\n            <li><a href=\"/556d5103a27286054615ccc0f2cd8db0.html\" class=\"cstore\">Admin Site</a></li>\n            <li><a href=\"/99bad02247f3eae83c76aa3fe371b715.html\" class=\"login\">Logout</a></li>\n        </ul>\n    </nav>\n    \n    <div class=\"welcome\" id=\"welcome\">\n        <h1 style=\"color: lightseagreen\">Available Products</h1>\n        <p id=\"output\"></p>\n        <ul id=\"products\"></ul>\n    </div>\n    <footer class=\"footercentered\">\n        <p class=\"footer-company-motto\">Sit back.StoreManger will handle it.</p>\n        <p class=\"footer-links\">\n            <a href=\"/eacf331f0ffc35d4b482f1d15a887d3b.html\">Home</a> ·\n            <a href=\"#\">Blog</a>   \n        </p>\n        <p class=\"footer-company-name\">Store Manager.inc &copy; 2018</p>\n    </footer>\n    <script type=\"text/javascript\" src=\"/1355ab0ba902d64bd4832495ccf71fea.js\"></script>\n</body>\n\n</html>\n"},"hash":"72da5c655f91519165b88dad10352fe0","cacheData":{"env":{}}}
