+++ UI/html/addattendant.html
@@ -19,7 +19,7 @@
             <li><a href="sales.html">Sales</a></li>
             <li><a href="#" class="active">Add User</a></li>
             <li><a href="products.html">Attendant Site</a></li>
-            <li><a href="./login.html" class="login">Logout</a></li>
+            <li><a onclick="logout()" class="login">Logout</a></li>
         </ul>
     </nav>
      <div class="welcome">
@@ -39,6 +39,7 @@
         </p>
         <p class="footer-company-name">Store Manager.inc &copy; 2018</p>
     </footer>
-     <script type="text/javascript" src="../static/js/new_user.js"></script>
+    <script type="text/javascript" src="../static/js/new_user.js"></script>
+    <script type="text/javascript" src="../static/js/logout.js"></script>
 </body>
 </html>
