+++ UI/html/attendant-sales.html
@@ -176,6 +176,5 @@
     </footer>
 
     <script type="text/javascript" src="../static/js/script.js"></script>
-    <script type="text/javascript" src="../static/js/dts.js"></script>
 </body>
 </html>
