+++ UI/html/login.html
@@ -41,7 +41,6 @@
         </p>
         <p class="footer-company-name">Store Manager.inc &copy; 2018</p>
     </footer>
-    <script type="text/javascript" src="../static/js/script.js"></script>
-    <script type="text/javascript" src="../static/js/auth.js"></script>
+    <script type="text/javascript" src="../static/js/login.js"></script>
 </body>
 </html>
