+++ UI/html/newstore.html
@@ -36,7 +36,7 @@
             <div class="col-75"><label >Confirm Password</label></div>
             <div class="col-75"><input id="confirmpassword" type="password"  placeholder="*******"></div>
             <p id="output"></p>
-            <a> <input type="submit" value="CreateStore"></a>
+            <a > <input type="submit" value="CreateStore"></a>
                 
         </div>
     </form>
@@ -48,7 +48,6 @@
         </p>
         <p class="footer-company-name">Store Manager.inc &copy; 2018</p>
     </footer>
-    <script type="text/javascript" src="../static/js/script.js"></script>
-    <script type="text/javascript" src="../static/js/auth.js"></script>
+    <script  src="../static/js/signup.js"></script>
 </body>
 </html>
