+++ UI/html/sales2.html
@@ -175,6 +175,5 @@
     </footer>
 
     <script type="text/javascript" src="../static/js/script.js"></script>
-    <script type="text/javascript" src="../static/js/dts.js"></script>
 </body>
 </html>
