+++ UI/html/stock.html
@@ -23,19 +23,18 @@
             <li><a href="./login.html" class="login">Logout</a></li>
         </ul>
     </nav>
-    
-        <div id="addtocart" class="modal" style="display:none">
-            <form id="newproduct">
-                <div class="col-75"><label >Product Name</label></div>
-                <div class="col-75"><input id="name" type="text"  placeholder="name of product.."></div>
-                <div class="col-75"><label >Number_Of_Items</label></div>
-                <div class="col-75"><input id="number" type="text"  placeholder="20 "></div>
-                <div class="col-75"><label >Price KSH</label></div>
-                <div class="col-75"><input id="price" type="text"  placeholder="200"></div>
-                <p id="output"></p>
-                <a href="#" ><input   type="submit" value="Done"><br></a>
-            <form>
-        </div>
+    <div>
+        <form id="newproduct"  class="modal" style="display:none">
+            <div class="col-75"><label >Product Name</label></div>
+            <div class="col-75"><input id="name" type="text"  placeholder="name of product.."></div>
+            <div class="col-75"><label >Number_Of_Items</label></div>
+            <div class="col-75"><input id="number" type="text"  placeholder="20 "></div>
+            <div class="col-75"><label >Price KSH</label></div>
+            <div class="col-75"><input id="price" type="text"  placeholder="200"></div>
+            <p id="output"></p>
+            <a href="#" ><input   type="submit" value="Done"><br></a>
+        <form>
+    </div>
     
     <div class="pcontainer">
         <h1>Available Products</h1>
@@ -69,7 +68,7 @@
         <p class="footer-company-name">Store Manager.inc &copy; 2018</p>
     </footer>
     <script type="text/javascript" src="../static/js/script.js"></script>
-    <script type="text/javascript" src="../static/js/admin_products.js"></script>
+    <script type="text/javascript" src="../static/js/addproduct.js"></script>
 </body>
 
 </html>
