+++ UI/static/css/products.css
@@ -52,7 +52,7 @@
 .welcome ul {
     margin-top: 20px;
     display: grid;
-    grid-template-columns: repeat(4, 1fr);
+    grid-template-columns: repeat(3, 1fr);
     grid-gap: 10px;
  
   }
