+++ UI/static/js/logout.js
@@ -1,0 +1,6 @@
+//logout function
+
+function logout (){
+    localStorage.clear();
+    window.location.href = "index.html";
+}
